pyproject.toml
[project]
name = "p1-reader"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomqtt>=2.3.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "ruff>=0.7.1",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -v"
log_level = "INFO"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
